Added camera to follow player.
